Added Reset function so user can create another city from the menu. Moved Car, Mesh, Entity, Deco & Building to Objective-C objects from C++.

+++ PixelCity/Car.h
@@ -1,7 +1,7 @@
 
 void   CarAdd();
 void   CarClear();
-int    CarCount();
+size_t CarCount();
 void   CarRender();
 void   CarUpdate();
